# 🚀 Major System Enhancement: Agent-Based Architecture with Performance Optimizations